requirements.txt
Django==1.9.6
django-contrib-comments==1.7.1
django-markdown-deux==1.0.5
djangorestframework==3.3.3
markdown2==2.3.1
pytz==2016.4
